+++ CHANGELOG.md
@@ -14,8 +14,9 @@
 ### Bugfixes
 - [#199](https://github.com/influxdata/kapacitor/issues/199): BREAKING: Various fixes for the Alerta integration.
     The `event` property has been removed from the Alerta node and is now set as the value of the alert ID.
+- [#232](https://github.com/influxdata/kapacitor/issues/232): Better error message for alert integrations. Better error message for VictorOps 404 response.
 
-## v0.10.1 [2013-02-08]
+## v0.10.1 [2016-02-08]
 
 ### Release Notes